requirements.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.2
numba==0.56.4
numba-progress==0.0.4
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.9.3
six==1.16.0
tqdm==4.64.1
